Build take_a_guess & parse_guess function
take_a_guess will be used similarly between generating code and guessing
code. parse_guess will be used to parse the user's guess. The function
asks player to select four colors out of the six available colors.
Anything that isn't a letter is removed from the guess and then
converted to uppercase. This is then tested against a regular expression
to ensure that the user's input is valid. As long as the only letters in
the input are the six colors, the function will except it. This means
that the user can submit a guess of any length and the function will
accept it as long as only the six colors are used. The parse_guess takes
the user's guess and splits it into an array of four elements. Only the
first four letters of the string are considered.