Build function comparing guess to code
Added three methods: one to check for perfect matches, one to check for
no matches, and then a third to build a clue. The clue is built by
adding the perfect matches to the no matches, and if the length of that
is less than 4, it adds a marker indicating a partial match.