Fix bug where clue was incorrectly adding 'W'
There was a bug where if the clue contained one of a letter, and your
guess had two of a letter with one of the letters being in the right
position. The presence of the second one would cause there to be a 'W'
in the clue creating nonsensical clues like 'BBBW'. This isn't possible
because you can't have three in the correct position and one that's the
right color but in the wrong position.