Add method for computer to guess & logic for turns
Added some logic in round to make it figure out whose turn it is
and based on the result either make the computer or player take a guess.
The method for the computer to take a guess is random unless it guesses
a perfect match. Then that perfect match gets added permanently to it's
guess. Still need to add logic for partial matches.